websocket

+++ Master/CashFlowAPP/BLL/Services/AdminSide/UsersService.cs
@@ -5,6 +5,7 @@ using Common.Enum;
 using Common.Model.AdminSide;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
+using System.Linq;
 
 namespace BLL.Services.AdminSide
 {
@@ -86,13 +87,14 @@ namespace BLL.Services.AdminSide
                         user = user.Where(x => x.Status == Status);
                     }
                 }
+           
 
                 var Data = user
                     // 後端分頁
                     // 省略幾筆 ( 頁數 * 每頁幾筆 )
-                    .Skip((Req.PageIndex - 1) * Req.PageSize)
+                    .Skip(((int)Req.PageIndex - 1) * (int)Req.PageSize)
                     // 取得幾筆
-                    .Take(Req.PageSize)
+                    .Take((int)Req.PageSize)
                     .ToList();
 
                 Res.Data = Data;
